update package name in test directory
